+++ src/ChannelConf.h
@@ -14,7 +14,7 @@ namespace CML {
     uint32_t duration = 0;  // Unit us.
     float area = 0; // Unit mm^2.
     float burst_frac = 1; // Fraction of 1/burst_slow_freq to stimulate for.
-    uint32_t burst_slow_freq = 0; // Unit Hz.  Slower envelope freq of bursts.
+    float burst_slow_freq = 0; // Unit Hz.  Slower envelope freq of bursts.
   };
 
   class BipolarPair {
